+++ locacao-service/src/main/java/com/example/locacao_service/dto/carrinho/CarrinhoRequestDto.java
@@ -13,5 +13,5 @@ import java.time.LocalDateTime;
 public class CarrinhoRequestDto {
 
     @Positive
-    private BigInteger pessoaId;
+    private Long pessoaId;
 }
